fix nllloss 2d bug
