+++ oneflow/python/test/modules/test_nllloss.py
@@ -118,41 +118,41 @@ class TestNLLLossModule(flow.unittest.TestCase):
         np_out = nll_loss_1d(input.numpy(), target.numpy(), reduction="sum")
         test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
 
-    # def test_nllloss_segmentation_none(test_case):
-    #     x = np.array(
-    #         [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
-    #     ).astype(np.float32)
-    #     input = flow.Tensor(x, dtype=flow.float32)
-    #     y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
-    #     target = flow.Tensor(y, dtype=flow.int64)
-    #     nll_loss = flow.nn.NLLLoss()
-    #     of_out = nll_loss(input, target)
-    #     np_out = nll_loss_2d(input.numpy(), target.numpy())
-    #     test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
+    def test_nllloss_segmentation_none(test_case):
+        x = np.array(
+            [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
+        ).astype(np.float32)
+        input = flow.Tensor(x, dtype=flow.float32)
+        y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
+        target = flow.Tensor(y, dtype=flow.int64)
+        nll_loss = flow.nn.NLLLoss()
+        of_out = nll_loss(input, target)
+        np_out = nll_loss_2d(input.numpy(), target.numpy())
+        test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
 
-    # def test_nllloss_segmentation_mean(test_case):
-    #     x = np.array(
-    #         [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
-    #     ).astype(np.float32)
-    #     input = flow.Tensor(x, dtype=flow.float32)
-    #     y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
-    #     target = flow.Tensor(y, dtype=flow.int64)
-    #     nll_loss = flow.nn.NLLLoss(reduction="mean")
-    #     of_out = nll_loss(input, target)
-    #     np_out = nll_loss_2d(input.numpy(), target.numpy(), reduction="mean")
-    #     test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
+    def test_nllloss_segmentation_mean(test_case):
+        x = np.array(
+            [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
+        ).astype(np.float32)
+        input = flow.Tensor(x, dtype=flow.float32)
+        y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
+        target = flow.Tensor(y, dtype=flow.int64)
+        nll_loss = flow.nn.NLLLoss(reduction="mean")
+        of_out = nll_loss(input, target)
+        np_out = nll_loss_2d(input.numpy(), target.numpy(), reduction="mean")
+        test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
 
-    # def test_nllloss_segmentation_sum(test_case):
-    #     x = np.array(
-    #         [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
-    #     ).astype(np.float32)
-    #     input = flow.Tensor(x, dtype=flow.float32)
-    #     y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
-    #     target = flow.Tensor(y, dtype=flow.int64)
-    #     nll_loss = flow.nn.NLLLoss(reduction="sum")
-    #     of_out = nll_loss(input, target)
-    #     np_out = nll_loss_2d(input.numpy(), target.numpy(), reduction="sum")
-    #     test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
+    def test_nllloss_segmentation_sum(test_case):
+        x = np.array(
+            [[[[0.12, 0.36], [0.22, 0.66]], [[0.13, 0.34], [0.52, -0.96]]]]
+        ).astype(np.float32)
+        input = flow.Tensor(x, dtype=flow.float32)
+        y = np.array([[[1, 0], [0, 1]]]).astype(np.int)
+        target = flow.Tensor(y, dtype=flow.int64)
+        nll_loss = flow.nn.NLLLoss(reduction="sum")
+        of_out = nll_loss(input, target)
+        np_out = nll_loss_2d(input.numpy(), target.numpy(), reduction="sum")
+        test_case.assertTrue(np.allclose(of_out.numpy(), np_out))
 
 
 if __name__ == "__main__":
